requirements.txt
bcrypt==4.3.0
email_validator==2.2.0
fastapi==0.115.12
psycopg2-binary==2.9.10
pytest==8.3.5
python-jose==3.4.0
python-multipart==0.0.20
SQLAlchemy==2.0.41
uvicorn==0.34.2
